fix warning for compile with v143 vs2022

+++ src/shadereditor/editorcommon.cpp
@@ -2252,5 +2252,5 @@ void BuildUserFunctionName( const char *pFunctionName, const CUtlVector< __funcP
 
 void BuildDefaultFunctionName( char *pszDest, int buflen, CBaseNode *pNode )
 {
-	Q_snprintf( pszDest, buflen, "UserFunction_%u", pNode->GetUniqueIndex() );
+	Q_snprintf( pszDest, buflen, "UserFunction_%u", (int) pNode->GetUniqueIndex() );
 }
